// src/enums/barcode-type-name.ts
export enum BarcodeTypeName {
  Code11 = 'Code11',
  Interleaved25 = 'Interleaved25',
  Code39 = 'Code39',
  PlanetCode = 'PlanetCode',
  PDF417 = 'PDF417',
  EAN8 = 'EAN8',
  UPCE = 'UPCE',
  Code93 = 'Code93',
  Code128 = 'Code128',
  EAN13 = 'EAN13',
  Industrial25 = 'Industrial25',
  Standard25 = 'Standard25',
  ANSICodabar = 'ANSICodabar',
  Logmars = 'Logmars',
  MSI = 'MSI',
  Plessey = 'Plessey',
  QRCode = 'QRCode',
  DataMatrix = 'DataMatrix',
  PostNet = 'PostNet',
}